again

+++ main.js
@@ -680,7 +680,7 @@ function canAffordBuilding(what, take, buildCostString, isEquipment){
 	for (var costItem in toBuy.cost) {
 		var color = "green";
 		var price = 0;
-		price = parseInt(getBuildingItemPrice(toBuy, costItem, isEquipment))
+		price = parseFloat(getBuildingItemPrice(toBuy, costItem, isEquipment))
 		if (price > game.resources[costItem].owned || !(isFinite(price))) {
 			if (buildCostString) color = "red";
 			else return false;
